Rename pool_data module to pools
